+++ ts/interpolations/attributes/processNameValueAttribute.function.ts
@@ -2,13 +2,11 @@ import { specialAttribute } from './specialAttribute.js'
 import { HowToSet } from './howToSetInputValue.function.js'
 import { TagGlobal } from '../../tag/getTemplaterResult.function.js'
 import { processTagCallbackFun } from '../../render/attributes/processAttribute.function.js'
-import { SpecialAction, SpecialDefinition } from '../../render/attributes/Special.types.js'
+import { SpecialDefinition } from '../../render/attributes/Special.types.js'
 import { ContextItem } from '../../tag/ContextItem.type.js'
 import { AnySupport } from '../../tag/AnySupport.type.js'
 import { BasicTypes } from '../../tag/ValueTypes.enum.js'
 import type { TagCounts } from '../../tag/TagCounts.type.js'
-
-const actions = ['init','destroy'] // oninit ondestroy
 
 export function processDynamicNameValueAttribute(
   attrName: string,
@@ -24,18 +22,6 @@ export function processDynamicNameValueAttribute(
   contextItem.howToSet = howToSet
 
   if(typeof(value) === BasicTypes.function ) {
-    if (isSpecial && actions.includes(attrName)) {
-      specialAttribute(
-        attrName,
-        value,
-        element,
-        attrName as SpecialAction,
-        support,
-        counts,
-      )
-      return
-    }
-  
     return processTagCallbackFun(
       contextItem,
       value,
@@ -74,8 +60,6 @@ export function processNonDynamicAttr(
       value,
       element,
       isSpecial,
-      support,
-      counts,
     )
   }