+++ ts/interpolations/attributes/renderTagArray.function.ts
@@ -13,6 +13,7 @@ export function renderTagUpdateArray(
 
   --painting.locks
 
+
   paint()
 }