+++ ts/tag/createHtmlSupport.function.ts
@@ -4,7 +4,7 @@ import { Subject } from '../subject/Subject.class.js'
 import { Props } from '../Props.js'
 import { BaseSupport } from './BaseSupport.type.js'
 import { AnySupport } from './AnySupport.type.js'
-import { ContextItem, SupportContextItem } from '../index.js'
+import { ContextItem, getNewGlobal, SupportContextItem } from '../index.js'
 
 export type PropsConfig = {
   latest: Props // new props NOT cloned props
@@ -37,11 +37,13 @@ export function getBaseSupport(
     appSupport: undefined as unknown as AnySupport,
   } as BaseSupport
   
-  // baseSupport.appSupport = baseSupport
 
+  // const global = context.global || getNewGlobal(context)
   const global = context.global
   global.blocked = []
   global.destroy$ = new Subject<void>()
+  // global.oldest = global.oldest || baseSupport as AnySupport
+  // global.oldest = baseSupport as AnySupport
 
   return baseSupport
 }